+++ docs/release-notes.rst
@@ -5,6 +5,8 @@ Future Release
 --------------
 
     * Add Redis 4 compatability fix to CLUSTER NODES command (See issue #217)
+    * Fixed bug with command "CLUSTER GETKEYSINSLOT" that was throwing exceptions
+    * Added new methods get_keys_in_slot() to client
 
 1.3.4 (Mar 5, 2017)
 -------------------